requirements.txt
serpapi
openai
python-dotenv
requests